+++ latexml_package/src/package.rs
@@ -124,6 +124,7 @@ pub mod crop_sty;
 pub mod cropmark_sty;
 pub mod csquotes_sty;
 pub mod ctable_sty;
+pub mod currfile_sty;
 pub mod datatool_base_sty;
 pub mod dcolumn_sty;
 pub mod delarray_sty;